requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
python-dotenv
gunicorn